requirements.txt
Django==2.2.27
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
stripe==4.1.0